backend/requirements.txt
Flask
flask-cors
tensorflow==2.16.1
Pillow
numpy
gunicorn